% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_PlanBsmooth.r
\name{ApplyPlanBsmooth}
\alias{ApplyPlanBsmooth}
\title{Applies the Plan B smooth approach.}
\usage{
ApplyPlanBsmooth(dat, od = ".\\\\", my.title = "", terminal.year = NA,
  nyears = 33, loess.span = NA, saveplots = FALSE, showplots = TRUE)
}
\arguments{
\item{dat}{data frame of Year and avg (the biomass index)}

\item{od}{output directory where plots are saved (default=working directory)}

\item{my.title}{title for time series plot with loess smooth (default = "")}

\item{terminal.year}{last year used in smooth (allows easy retro analysis) (default = NA = most recent)}

\item{nyears}{number of years to use in loess (default = 33)}

\item{loess.span}{proportion time series used in smoothing (default = NA, calculates span=9.9/nyears)}

\item{saveplots}{true/false flag to save output to od (default=FALSE)}

\item{showplots}{true/false flag to display plots in window (default=TRUE), must show plots to save}
}
\description{
Smooths data (Year and Biomass Index) using loess, applies log linear regression to most recent three years, and retransforms back to estimate multiplier for catch advice.
}
