% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_raw.r
\name{ReadRaw}
\alias{ReadRaw}
\title{Get raw data and prepare for Plan B smooth}
\usage{
ReadRaw(data.dir, data.file.name, od = ".\\\\", standardize = FALSE,
  usesurvey = NA, saveplots = FALSE)
}
\arguments{
\item{data.dir}{directory with csv file}

\item{data.file.name}{name of csv file (without the .csv extension)}

\item{od}{directory for output (default=working directory)}

\item{standardize}{true/false flag to divide by mean of time series (default=FALSE)}

\item{usesurvey}{vector of true/false to select surveys to average (default=NA which means use all)}

\item{saveplots}{true/false flag to save output to od (default=FALSE)}
}
\description{
Reads in csv file with Year in first column and as many survey indices as desired, standardizes series if desired, computes mean of all or user selected surveys, outputs csv file with Year and Mean Index.
}
