% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_retro.r
\name{RunRetro}
\alias{RunRetro}
\title{Wrapper function to conduct retrospective analysis.}
\usage{
RunRetro(dat, od = ".\\\\", my.title = "", terminal.year = NA,
  nyears = 33, loess.span = NA, npeels = 7, saveretroplot = FALSE)
}
\arguments{
\item{dat}{data frame of Year and avg (the biomass index)}

\item{od}{output directory where plots are saved (default=working directory)}

\item{my.title}{title for time series plot with loess smooth (default = "")}

\item{terminal.year}{last year used in smooth (allows easy retro analysis) (default = NA = most recent)}

\item{nyears}{number of years to use in loess (default = 33)}

\item{loess.span}{proportion time series used in smoothing (default = NA, calculates span=9.9/nyears)}

\item{npeels}{how many years to remove in retrospective analysis}

\item{saveretroplot}{true/false flag to save retrospective plot to od (default=FALSE)}
}
\description{
Conducts standard retrospective analysis on predicted biomass index, inlcuding reporting Mohn's rho. Also examines retrospective analysis of multiplier (inclusing years prior to terminal year).
}
